#188 - :construction: WIP - Seleção de tela para carregar a partir dos tipos de ocorrência
