+++ src/pages/FaleConoscoScreen/tiposDeOcorrencia.js
@@ -1,0 +1,5 @@
+const ALERTA_FALTA_EPI = 'Alerta de falta de EPI';
+const RELATAR_SUGESTAO = 'Relatar sugestão (iSUS)';
+const RELATAR_PROBLEMA = 'Relatar problema (iSUS)';
+
+export { ALERTA_FALTA_EPI, RELATAR_SUGESTAO, RELATAR_PROBLEMA };
